* rename visualisation -> drawing * move hist-matching * add test hist-matching
