+++ birl/utilities/drawing.py
